requirements.txt
bitstruct
python-dotenv